+++ activerecord/lib/active_record/connection_adapters/sqlite3_adapter.rb
@@ -375,10 +375,12 @@ module ActiveRecord
           row['name']
         end
       end
+      alias data_sources tables
 
       def table_exists?(table_name)
         table_name && tables(nil, table_name).any?
       end
+      alias data_source_exists? table_exists?
 
       # Returns an array of +Column+ objects for the table specified by +table_name+.
       def columns(table_name) #:nodoc: