+++ activerecord/test/cases/adapter_test.rb
@@ -38,6 +38,21 @@ module ActiveRecord
       assert !@connection.table_exists?(nil)
     end
 
+    def test_data_sources
+      data_sources = @connection.data_sources
+      assert data_sources.include?("accounts")
+      assert data_sources.include?("authors")
+      assert data_sources.include?("tasks")
+      assert data_sources.include?("topics")
+    end
+
+    def test_data_source_exists?
+      assert @connection.data_source_exists?("accounts")
+      assert @connection.data_source_exists?(:accounts)
+      assert_not @connection.data_source_exists?("nonexistingtable")
+      assert_not @connection.data_source_exists?(nil)
+    end
+
     def test_indexes
       idx_name = "accounts_idx"