+++ activerecord/CHANGELOG.md
@@ -1,3 +1,9 @@
+*   Introduce `connection.data_sources` and `connection.data_source_exists?`.
+    These methods determine what relations can be used to back Active Record
+    models (usually tables and views).
+
+    *Yves Senn*, *Matthew Draper*
+
 *   No longer pass deprecated option `-i` to `pg_dump`.
 
     *Paul Sadauskas*